services/backend/pyproject.toml
[tool.poetry]
name = "journal-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "*"
fastapi = "*"
tortoise-orm = "*"
aerich = "*"
asyncpg = "*"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "*"
python-multipart = "*"



[tool.poetry.group.dev.dependencies]
httpx = "*"
pytest = "*"
ruff = "*"
mypy = "*"

[tool.aerich]
tortoise_orm = "src.core.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
